+++ example/findPerformances.js
@@ -3,14 +3,14 @@
  *
  * @ignore
  */
-import * as COA from '../lib/index';
+
+const COA = require('../');
 
 COA.services.master.schedule({
     theaterCode: '118',
     begin: '20170411',
     end: '20170411'
 }).then((performances) => {
-    // tslint:disable-next-line:no-console
     console.log(performances);
 }).catch((err) => {
     console.error(err.message);