+++ example/stateReserve.js
@@ -3,14 +3,14 @@
  *
  * @ignore
  */
-import * as COA from '../lib/index';
+
+const COA = require('../');
 
 COA.services.reserve.stateReserve({
     theaterCode: '118',
     reserveNum: 1339,
     telNum: '09012345678'
 }).then((result) => {
-    // tslint:disable-next-line:no-console
     console.log(result);
 }).catch((err) => {
     console.error(err.message);